perbaikan crud dan response deliv dan sale

+++ backend/app/schemas/input_models/deliveries_input_models.py
@@ -22,7 +22,7 @@ class DeliveryResponse(BaseModel):
     id: int
     code: str
     date: Optional[str] = None
-    quantity: Optional[int] = None
+    quantity: Optional[float] = None
     sale_id: Optional[int] = None
     return_id: Optional[int] = None
     roll: Optional[float] = None
